refactor(graphql-client): вынести создание кэша в отдельный файл
Jira issue: VEGA-911
