+++ src/services/graphql-client/graphql-client.ts
@@ -4,16 +4,15 @@ import {
   createHttpLink as createApolloHttpLink,
   from,
   HttpOptions,
-  InMemoryCache,
   NormalizedCacheObject,
   Observable,
   ServerParseError,
-  StoreObject,
 } from '@apollo/client';
 import { ErrorResponse, onError } from '@apollo/client/link/error';
 
 import { Identity } from '../identity/identity';
 
+import { createCache } from './cache';
 import { ProjectDiffResolverLink } from './project-diff-resolver';
 
 export type GraphQLClient = ApolloClient<NormalizedCacheObject>;
@@ -37,51 +36,6 @@ export type GraphQLClientConfig = {
 
 type ResponseLinkConfig = {
   handleError: ErrorHandler;
-};
-
-export const getDataIdFromObject = (
-  obj: Readonly<StoreObject>,
-  context: { keyObject?: Record<string, unknown> },
-): string | undefined => {
-  // eslint-disable-next-line no-underscore-dangle
-  const { id, vid, _id, __typename } = obj;
-
-  if (typeof __typename !== 'string') {
-    return undefined;
-  }
-
-  let resultId = id;
-
-  // istanbul ignore else
-  if (context !== undefined) {
-    if (id !== undefined) {
-      context.keyObject = { id };
-    } else if (vid !== undefined) {
-      context.keyObject = { vid };
-    } else if (_id !== undefined) {
-      context.keyObject = { _id };
-    } else {
-      context.keyObject = undefined;
-    }
-  }
-
-  if (_id !== undefined) {
-    resultId = _id;
-  }
-
-  if (vid !== undefined) {
-    resultId = vid;
-  }
-
-  if (resultId === undefined || resultId === null) {
-    return undefined;
-  }
-
-  return `${__typename}:${
-    typeof resultId === 'string' || typeof resultId === 'number'
-      ? resultId
-      : JSON.stringify(resultId)
-  }`;
 };
 
 export function normalizeUri(uri: string): string {
@@ -214,14 +168,7 @@ export function createGraphqlClient(config: GraphQLClientConfig): GraphQLClient 
   const { uri, identity, onError: handleError, fetch } = config;
   return new ApolloClient({
     connectToDevTools: process.env.VEGA_ENV === 'development',
-    cache: new InMemoryCache({
-      dataIdFromObject: getDataIdFromObject,
-      typePolicies: {
-        Project: {
-          keyFields: ['vid'],
-        },
-      },
-    }),
+    cache: createCache(),
     link: from([
       createErrorLink({ handleError }),
       createSwitchUriLink(uri),
